+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Implement Bugzilla SRT notes builder in Bugzilla Backwards Sync (OSIDB-384)
+- Implement validation for flaw without affect (OSIDB-353)
 
 ### Changed
 - Change logging of celery and django to filesystem (OSIDB-418)